BE/flask/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
distro==1.8.0
fastapi==0.104.1
fonttools==4.45.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
openai==1.3.3
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pydantic==2.5.1
pydantic_core==2.14.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
uvicorn
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
